perf: cut per-emit CSV rewrite and per-record allocations
Debounce Output.csv behind a dirty flag, share one RocksDB cache, drop SHA-256 ids and regex compiles, pipeline Kafka acks per poll, and aggregate with long instead of BigInteger.

+++ backend/src/main/java/com/example/amn/Transaction.java
@@ -1,6 +1,4 @@
 package com.example.amn;
 
-import java.math.BigInteger;
-
 public record Transaction(String id, String clientInformation, String productInformation,
-                          BigInteger delta, long sourcePosition) {}
+                          long delta, long sourcePosition) {}
